+++ src/main/net-constraints-nunit3/EvaluateXPathConstraint.cs
@@ -108,7 +108,7 @@ namespace Org.XmlUnit.Constraints
         /// <inheritdoc/>
         public override ConstraintResult ApplyTo<TActual>(TActual actual)
         {
-            ConstraintResult nested = valueConstraint.ApplyTo(XPathEvaluate(actual));
+            ConstraintResult nested = valueConstraint.Resolve().ApplyTo(XPathEvaluate(actual));
             Description = "XML with XPath " + xPath + " evaluated to " + nested.ActualValue;
             return new HasXPathConstraintResult(this, actual, nested);
         }