+++ src/tests/net-constraints-nunit3/EvaluateXPathConstraintTest.cs
@@ -133,5 +133,22 @@ namespace Org.XmlUnit.Constraints {
                         .With.Property("Message").Contains("3"));
         }
 
+        [Test]
+        public void CreatesAUsefulMessageWhenFailingCombinedWithNotOnTheInside() {
+            string xml = "<?xml version=\"1.0\" encoding=\"UTF-8\"?>" +
+                "<fruits>" +
+                "<fruit name=\"apple\"/>" +
+                "<fruit name=\"orange\"/>" +
+                "<fruit name=\"banana\"/>" +
+                "</fruits>";
+            Assert.That(() =>
+                        Assert.That(xml, EvaluateXPathConstraint.HasXPath("count(//fruits/fruit)",
+                                                                          !Is.EqualTo("3"))),
+                        Throws.TypeOf<AssertionException>()
+                        .With.Property("Message").Contains("XML with XPath count(//fruits/fruit)")
+                        .With.Property("Message").Contains("not")
+                        .With.Property("Message").Contains("3"));
+        }
+
     }
 }